#3155 Rename packages
#3155 Create special pharma import product table
