+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/ad/table/process/AD_ImpFormat_Row_Create_Based_OnTable.java
@@ -1,4 +1,4 @@
-package org.adempiere.ad.table.api.impl;
+package org.adempiere.ad.table.process;
 
 import static org.adempiere.model.InterfaceWrapperHelper.newInstance;
 import static org.adempiere.model.InterfaceWrapperHelper.save;
