+++ de.metas.adempiere.adempiere/migration/src/main/sql/postgresql/system/10-de.metas.adempiere/5480071_sys_gh3155_AddCreateImportFormatRowsBasedOnTableColumns.sql
@@ -16,3 +16,7 @@ INSERT INTO AD_Table_Process (AD_Client_ID,AD_Org_ID,AD_Process_ID,AD_Table_ID,C
 ;
 
 
+-- 2017-12-19T11:24:33.629
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Process SET Classname='org.adempiere.ad.table.process.AD_ImpFormat_Row_Create_Based_OnTable',Updated=TO_TIMESTAMP('2017-12-19 11:24:33','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Process_ID=540896
+;
