+++ de.metas.vertical.pharma/src/main/java/de/metas/impexp/process/ImportPharmaProduct.java
@@ -1,4 +1,4 @@
-package org.adempiere.impexp.process;
+package de.metas.impexp.process;
 
 import org.compiere.process.AbstractImportJavaProcess;
